+++ apps/docs/content/components/accordion/index.ts
@@ -2,7 +2,10 @@ import usage from "./usage";
 import subtitle from "./subtitle";
 import multiple from "./multiple";
 import compact from "./compact";
-import variants from "./variants";
+import lightVariant from "./light-variant";
+import shadowVariant from "./shadow-variant";
+import splittedVariant from "./splitted-variant";
+import borderedVariant from "./bordered-variant";
 import defaultExpandedKeys from "./default-expanded-keys";
 import disabledKeys from "./disabled-keys";
 import startContent from "./start-content";
@@ -17,7 +20,10 @@ export const accordionContent = {
   subtitle,
   multiple,
   compact,
-  variants,
+  lightVariant,
+  shadowVariant,
+  splittedVariant,
+  borderedVariant,
   defaultExpandedKeys,
   disabledKeys,
   startContent,