+++ apps/docs/content/components/tabs/controlled.ts
@@ -13,21 +13,21 @@ export default function App() {
         <Tab key="photos" title="Photos">
           <Card>
             <CardBody>
-              List of photos
+              Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do eiusmod tempor incididunt ut labore et dolore magna aliqua. Ut enim ad minim veniam, quis nostrud exercitation ullamco laboris nisi ut aliquip ex ea commodo consequat.
             </CardBody>
           </Card>  
         </Tab>
         <Tab key="music" title="Music">
           <Card>
             <CardBody>
-              List of music
+              Ut enim ad minim veniam, quis nostrud exercitation ullamco laboris nisi ut aliquip ex ea commodo consequat. Duis aute irure dolor in reprehenderit in voluptate velit esse cillum dolore eu fugiat nulla pariatur.
             </CardBody>
           </Card>  
         </Tab>
         <Tab key="videos" title="Videos">
           <Card>
             <CardBody>
-              List of videos
+              Excepteur sint occaecat cupidatat non proident, sunt in culpa qui officia deserunt mollit anim id est laborum.
             </CardBody>
           </Card>  
         </Tab>