+++ packages/components/accordion/stories/accordion.stories.tsx
@@ -246,7 +246,7 @@ const CustomWithClassNamesTemplate: ComponentStory<typeof Accordion> = (args: Ac
   const itemStyles: AccordionItemProps["classNames"] = {
     base: "py-0 w-full",
     title: "font-normal text-base",
-    trigger: "px-2 py-0 hover:bg-default-100 rounded-lg h-14 flex items-center",
+    trigger: "px-2 py-0 data-[hover=true]:bg-default-100 rounded-lg h-14 flex items-center",
     indicator: "text-base",
     content: "text-sm px-2",
   };